pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "google-genai>=1.29.0",
    "pydantic>=2.11.7",
    "python-telegram-bot>=22.3",
    "sift-stack-py>=0.8.2",
    "telegram>=0.0.1",
]
